requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
asgiref==3.5.2
    # via django
django==4.1.1
    # via
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.0
    # via -r requirements.in
pyjwt==2.5.0
    # via djangorestframework-simplejwt
pytz==2022.2.1
    # via djangorestframework
sqlparse==0.4.2
    # via django
